--- xprof/aggreg.cpp
@@ -1,4 +1,0 @@
-#include <metababel/metababel.h>
-
-void btx_register_usr_callbacks(void *btx_handle) {
-}
